+++ enterprise/cmd/frontend/auth/gitlaboauth/config.go
@@ -1,11 +1,12 @@
-package githuboauth
+package gitlaboauth
 
 import (
+	"fmt"
+	"net/url"
 	"os"
 	"strings"
 	"sync"
 
-	"github.com/dghubble/gologin"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/external/auth"
 	"github.com/sourcegraph/sourcegraph/pkg/conf"
 	"github.com/sourcegraph/sourcegraph/schema"
@@ -26,7 +27,7 @@ func init() {
 
 	var (
 		mu  sync.Mutex
-		cur = map[schema.GitHubAuthProvider]auth.Provider{} // tracks current mapping of valid config to auth.Provider
+		cur = map[schema.GitLabAuthProvider]auth.Provider{} // tracks current mapping of valid config to auth.Provider
 	)
 
 	go func() {
@@ -34,8 +35,8 @@ func init() {
 			mu.Lock()
 			defer mu.Unlock()
 
-			if !conf.Get().ExperimentalFeatures.GithubAuth {
-				new := map[schema.GitHubAuthProvider]auth.Provider{}
+			if !conf.Get().ExperimentalFeatures.GitlabAuth {
+				new := map[schema.GitLabAuthProvider]auth.Provider{}
 				updates := make(map[auth.Provider]bool)
 				for c, p := range cur {
 					if _, ok := new[c]; !ok {
@@ -48,7 +49,7 @@ func init() {
 				return
 			}
 
-			log15.Info("Reloading changed GitHub OAuth authentication provider configuration.")
+			log15.Info("Reloading changed GitLab OAuth authentication provider configuration.")
 
 			new, _ := parseConfig(conf.Get())
 			updates := make(map[auth.Provider]bool)
@@ -73,31 +74,30 @@ func init() {
 	})
 }
 
-func parseConfig(cfg *schema.SiteConfiguration) (providers map[schema.GitHubAuthProvider]auth.Provider, problems []string) {
-	providers = make(map[schema.GitHubAuthProvider]auth.Provider)
+func parseConfig(cfg *schema.SiteConfiguration) (providers map[schema.GitLabAuthProvider]auth.Provider, problems []string) {
+	providers = make(map[schema.GitLabAuthProvider]auth.Provider)
 	for _, pr := range cfg.AuthProviders {
-		if pr.Github == nil {
+		if pr.Gitlab == nil {
 			continue
 		}
 
-		provider, providerProblems := parseProvider(pr.Github, pr)
+		if cfg.ExternalURL == "" {
+			problems = append(problems, "`externalURL` was empty and it is needed to determine the OAuth callback URL.")
+			continue
+		}
+		externalURL, err := url.Parse(cfg.ExternalURL)
+		if err != nil {
+			problems = append(problems, fmt.Sprintf("Could not parse `externalURL`, which is needed to determine the OAuth callback URL."))
+			continue
+		}
+		callbackURL := *externalURL
+		callbackURL.Path = "/.auth/gitlab/callback"
+
+		provider, providerProblems := parseProvider(callbackURL.String(), pr.Gitlab, pr)
 		problems = append(problems, providerProblems...)
 		if provider != nil {
-			providers[*pr.Github] = provider
+			providers[*pr.Gitlab] = provider
 		}
 	}
 	return providers, problems
-}
-
-func getStateConfig() gologin.CookieConfig {
-	cfg := gologin.CookieConfig{
-		Name:     "github-state-cookie",
-		Path:     "/",
-		MaxAge:   120, // 120 seconds
-		HTTPOnly: true,
-	}
-	if conf.Get().TlsCert != "" {
-		cfg.Secure = true
-	}
-	return cfg
 }