+++ enterprise/cmd/frontend/auth/gitlaboauth/middleware.go
@@ -1,28 +1,28 @@
-package githuboauth
+package gitlaboauth
 
 import (
 	"net/http"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/external/auth"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/auth/oauth"
-	"github.com/sourcegraph/sourcegraph/pkg/extsvc/github"
+	"github.com/sourcegraph/sourcegraph/pkg/extsvc/gitlab"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
 
-const authPrefix = auth.AuthURLPrefix + "/github"
+const authPrefix = auth.AuthURLPrefix + "/gitlab"
 
 func init() {
 	oauth.AddIsOAuth(func(p schema.AuthProviders) bool {
-		return p.Github != nil
+		return p.Gitlab != nil
 	})
 }
 
 var Middleware = &auth.Middleware{
 	API: func(next http.Handler) http.Handler {
-		return ffHandler(oauth.NewHandler(github.ServiceType, authPrefix, true, next), next)
+		return ffHandler(oauth.NewHandler(gitlab.ServiceType, authPrefix, true, next), next)
 	},
 	App: func(next http.Handler) http.Handler {
-		return ffHandler(oauth.NewHandler(github.ServiceType, authPrefix, false, next), next)
+		return ffHandler(oauth.NewHandler(gitlab.ServiceType, authPrefix, false, next), next)
 	},
 }